Implement 64-bit integers and add_action command system
- Widen LPC integers from 32-bit to 64-bit (long)
  - Supports large values for XP, gold, timestamps without overflow
  - Matches modern LDMud/FluffOS behavior
  - Updated AST, Parser, Interpreter, ObjectInterpreter, and all efuns

- Add add_action() command system with three-tier resolution:
  - Protected commands (quit, score, etc.) - never overridable
  - Core commands (/cmds/) - overridable with AA_OVERRIDE flag
  - Custom actions via add_action() - context-sensitive commands
  - Includes query_verb(), notify_fail(), enable/disable_commands()

- Add type predicates: intp, stringp, objectp, pointerp/arrayp, mappingp, clonep
- Add utility efuns: allocate, copy, replace_string, trim
- Add say() efun for room communication
- Update documentation for integer size and new efuns

+++ src/Driver/Driver/Ast.cs
@@ -11,8 +11,9 @@ public abstract record Expression
 
 /// <summary>
 /// Integer literal: 42
+/// LPC integers are 64-bit to support large values (XP, gold, etc.)
 /// </summary>
-public record NumberLiteral(int Value) : Expression;
+public record NumberLiteral(long Value) : Expression;
 
 /// <summary>
 /// String literal: "hello"
